djangox/requirements.txt
asgiref==3.8.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==2024.10
cryptography==39.0.1
defusedxml==0.7.1
Django==4.2.9
django-allauth==65.0.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
dnspython==2.7.0
gunicorn==23.0.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.5.1
typing_extensions==4.9.0
urllib3==1.26.14
whitenoise==6.7.0
